[None][fix] Split TRT-LLM-only rope fusion out of standalone auto_deploy
Move FuseRopeIntoTrtllmAttention and its TRT-LLM-specific helpers from
rope.py into a new fuse_rope_into_trtllm_attention.py. The previous
top-level import of _TRTLLM_ROPE_INFO_KEY from trtllm_attention pulled
in tensorrt_llm runtime symbols at module-load time, which broke
rope.py for the standalone auto_deploy package and caused 22 rope and
qkv-fusion tests to fail in standalone mode.

The new file is silently skipped by library/__init__.py's import guard
when tensorrt_llm is unavailable, so all backend-agnostic rope
transforms in rope.py keep working in standalone mode.

Also split the two trtllm cache-insertion tests
(test_fuse_qkv_with_trtllm_cache_insertion,
 test_fuse_qkv_gqa_with_trtllm_cache_insertion) into
test_gemm_fusion_trtllm.py and add it to EXCLUDE_TEST_FILES.
